Cube color

+++ include/moteur/main.h
@@ -7,9 +7,9 @@
 
 typedef struct Color3f
 {
-	int r;
-	int g;
-	int b;
+	float r;
+	float g;
+	float b;
 } Color3f;
